feat: add CBOR-native ABI response types; remove serde_json
- Add AbiResponse, AbiAttribute, AbiEvent to abi module for CBOR-native
  serialization of execute/instantiate/external_event results
- Add AbiResultPayload to abi module (result as serde_bytes Vec<u8>, shared
  across all contracts)
- Fix IdbStateDump serde_as to use serde_with::Bytes for CBOR byte encoding
- Fix network_id: u8 -> u32 to handle chain IDs > 255
- Remove unused serde_json dependency

+++ macros/ownable-std-macros/src/lib.rs
@@ -315,7 +315,7 @@ pub fn ownables_instantiate_msg(metadata: TokenStream, input: TokenStream) -> To
             pub package: String,
             pub nft: Option<NFT>,
             pub ownable_type: Option<String>,
-            pub network_id: u8,
+            pub network_id: u32,
         }
     }
     .into();
